Demand integers in Tiny's operators and test the binary parser
Tiny's grammar has one numeric literal rule, integer, so add and subtract
wired NUMERIC for a float the language cannot express. They now wire
INTEGER. No behaviour changes: Tiny runs on DefaultTypeManager, which
types every expression I64, so the rule could never be violated.

BinarySemanticsParser gained real logic when the operand rules were
composed into it, and had no tests of its own. Adds 23, driving it through
a stub parser and a minimal type manager: rule conjunction, which rule
reports when several are violated, or-composed rules, same-family
promotion on either side, the mixed int/float rejection, the zero-divisor
value rule, and the two-error case where an operand error and a
follow-on promotion error are both reported.

+++ jcc-tiny/src/main/java/se/dykstrom/jcc/tiny/compiler/TinySemanticsParser.java
@@ -38,7 +38,7 @@ import java.util.Map;
 import java.util.Optional;
 
 import static se.dykstrom.jcc.common.error.Warning.UNUSED_VARIABLE;
-import static se.dykstrom.jcc.common.semantics.expression.OperandTypeRule.NUMERIC;
+import static se.dykstrom.jcc.common.semantics.expression.OperandTypeRule.INTEGER;
 
 /**
  * The semantics parser for the Tiny language.
@@ -64,11 +64,12 @@ public class TinySemanticsParser extends AbstractSemanticsParser<TypeManager> {
         statementComponents.put(WriteStatement.class, new WriteSemanticsParser<>(this));
 
         // Expressions
-        expressionComponents.put(AddExpression.class, new BinarySemanticsParser<>(this, "add", NUMERIC));
+        // The only numbers Tiny has are integers, so the operators demand integers
+        expressionComponents.put(AddExpression.class, new BinarySemanticsParser<>(this, "add", INTEGER));
         expressionComponents.put(IdentifierDerefExpression.class, new IdentifierDerefSemanticsParser<>(this, usageTracker));
         expressionComponents.put(IdentifierNameExpression.class, new IdentifierNameSemanticsParser<>(this, usageTracker));
         expressionComponents.put(IntegerLiteral.class, new IntegerSemanticsParser<>(this));
-        expressionComponents.put(SubExpression.class, new BinarySemanticsParser<>(this, "subtract", NUMERIC));
+        expressionComponents.put(SubExpression.class, new BinarySemanticsParser<>(this, "subtract", INTEGER));
     }
 
     @Override
